requirements.txt
asgiref==3.2.10
click==7.1.2
Django==3.0.5
djangorestframework==3.11.1
djongo==1.3.3
dnspython==2.0.0
h11==0.9.0
httptools==0.1.1
pymongo==3.11.0
python-dotenv==0.14.0
pytz==2020.1
sqlparse==0.2.4
uvicorn==0.11.8
uvloop==0.14.0
websockets==8.1
